+++ src/rust/engine/Cargo.toml
@@ -21,7 +21,6 @@ resolver = "2"
 members = [
   ".",
   "async_latch",
-  "async_semaphore",
   "async_value",
   "cache",
   "client",
@@ -62,7 +61,6 @@ members = [
 default-members = [
   ".",
   "async_latch",
-  "async_semaphore",
   "async_value",
   "cache",
   "client",
@@ -105,7 +103,6 @@ default = []
 
 [dependencies]
 async_latch = { path = "async_latch" }
-async_semaphore = { path = "async_semaphore" }
 # Pin async-trait due to https://github.com/dtolnay/async-trait/issues/144.
 async-trait = "=0.1.42"
 protos = { path = "protos" }
