--- src/rust/engine/async_semaphore/Cargo.toml
@@ -1,14 +1,0 @@
-[package]
-version = "0.0.1"
-edition = "2021"
-name = "async_semaphore"
-authors = [ "Pants Build <pantsbuild@gmail.com>" ]
-publish = false
-
-[dependencies]
-parking_lot = "0.11"
-tokio = { version = "1.4", features = ["sync"] }
-
-[dev-dependencies]
-futures = "0.3"
-tokio = { version = "1.4", features = ["rt", "macros", "time"] }
